update based on feedback on PR
