+++ Makefile
@@ -146,7 +146,7 @@ ctrl-manifests: controller-gen
 
 # Generate code
 ctrl-generate: controller-gen
-	$(CONTROLLER_GEN) object:headerFile="hack/boilerplate.go.txt" paths="./..."
+	$(CONTROLLER_GEN) object:headerFile="build-support/controller/boilerplate.go.txt" paths="./..."
 
 # find or download controller-gen
 # download controller-gen if necessary
