+++ PROJECT
@@ -1,3 +1,4 @@
+# this is a generated file used for operator sdk during code generation of CRDs, Controllers and webhooks
 domain: hashicorp.com
 layout: go.kubebuilder.io/v2
 repo: github.com/hashicorp/consul-k8s
