+++ build-support/controller/README.md
@@ -1,0 +1,9 @@
+# Consul-K8s hack GuideLines
+
+This document describes how you can use the scripts from [`hack`](.) directory 
+and gives a brief introduction and explanation of these scripts. 
+
+## Overview
+
+The [`hack`](.) directory contains many scripts that ensure continuous development of consul-k8s, 
+enhance the robustness of the code, improve development efficiency, etc.
